Move global types to global.d.ts
