--- frontend/src/types/auth/permission.types.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright © 2024 Hexastack. All rights reserved.
- *
- * Licensed under the GNU Affero General Public License v3.0 (AGPLv3) with the following additional terms:
- * 1. The name "Hexabot" is a trademark of Hexastack. You may not use this name in derivative works without express written permission.
- * 2. All derivative works must include clear attribution to the original creator and software, Hexastack and Hexabot, in a prominent location (e.g., in the software's "About" section, documentation, and README file).
- */
-
-import { EntityType } from "@/services/types";
-
-import { PermissionAction } from "../permission.types";
-import { IRole } from "../role.types";
-
-export interface IUserPermissions {
-  roles: IRole[];
-  permissions: Array<{
-    action: PermissionAction;
-    model: EntityType;
-    relation: string;
-  }>;
-}
